+++ astro.config.mjs
@@ -13,7 +13,7 @@ import remarkDirective from "remark-directive";
 import { remarkNote, addClassNames } from './src/plugins/markdown.custom'
 // Markdown 配置================
 import SITE_INFO from './src/config';
-import swup from '@swup/astro';
+
 // https://astro.build/config
 export default defineConfig({
 	site: SITE_INFO.Site,
@@ -21,25 +21,15 @@ export default defineConfig({
 	experimental: {
 		svgo: true,
 	},
-	integrations: [swup({
-		theme: false,
-		animationClass: "vh-animation-",
-		containers: [".main-inner", '.vh-header>.main'],
-		smoothScrolling: true,
-		progress: true,
-		cache: true,
-		preload: true,
-		accessibility: true,
-		updateHead: true,
-		updateBodyClass: false,
-		globalInstance: true
-	}),
-	Compress({ CSS: false, Image: false, Action: { Passed: async () => true } }),
-	sitemap({
-		changefreq: 'weekly', priority: 0.7, lastmod: new Date(),
-		// 处理末尾带 / 的 url
-		serialize: (item) => ({ ...item, url: item.url.endsWith('/') ? item.url.slice(0, -1) : item.url })
-	}), mdx({ extendMarkdownConfig: false })],
+	integrations: [
+		Compress({ CSS: false, Image: false, Action: { Passed: async () => true } }),
+		sitemap({
+			changefreq: 'weekly', priority: 0.7, lastmod: new Date(),
+			// 处理末尾带 / 的 url
+			serialize: (item) => ({ ...item, url: item.url.endsWith('/') ? item.url.slice(0, -1) : item.url })
+		}),
+		mdx({ extendMarkdownConfig: false })
+	],
 	markdown: {
 		remarkPlugins: [remarkMath, remarkDirective, remarkNote,],
 		rehypePlugins: [rehypeKatex, rehypeSlug, addClassNames],